refine copy handler
